Add Lib/PositionSanityChecker
- Refactor sanity checker code out of the Position model into a new
  PositionSanityChecker class.
- Update PositionSanitiyCheckController to use the new
  PostionSanityChecker class.
